// api/types/AppointmentResponse.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { AdminInfo } from "./AdminInfo.ts";
import type { DoctorInfo } from "./DoctorInfo.ts";
import type { UserInfo } from "./UserInfo.ts";

export const appointmentResponseTypeEnum = {
    "IN_PERSON": "IN_PERSON",
    "ONLINE": "ONLINE"
} as const;

export type AppointmentResponseTypeEnumKey = (typeof appointmentResponseTypeEnum)[keyof typeof appointmentResponseTypeEnum];

export const appointmentResponseStatusEnum = {
    "PENDING": "PENDING",
    "APPROVED": "APPROVED",
    "REJECTED": "REJECTED",
    "CANCELLED": "CANCELLED",
    "COMPLETED": "COMPLETED"
} as const;

export type AppointmentResponseStatusEnumKey = (typeof appointmentResponseStatusEnum)[keyof typeof appointmentResponseStatusEnum];

export type AppointmentResponse = {
    /**
     * @type integer | undefined, int64
    */
    id?: number;
    /**
     * @type string | undefined
    */
    type?: AppointmentResponseTypeEnumKey;
    /**
     * @type string | undefined
    */
    status?: AppointmentResponseStatusEnumKey;
    /**
     * @type string | undefined, date-time
    */
    appointmentDateTime?: string;
    /**
     * @type integer | undefined, int32
    */
    durationMinutes?: number;
    /**
     * @type object | undefined
    */
    user?: UserInfo;
    /**
     * @type object | undefined
    */
    doctor?: DoctorInfo;
    /**
     * @type object | undefined
    */
    assignedByAdmin?: AdminInfo;
    /**
     * @type string | undefined
    */
    userNotes?: string;
    /**
     * @type string | undefined
    */
    adminNotes?: string;
    /**
     * @type string | undefined
    */
    rejectionReason?: string;
    /**
     * @type string | undefined
    */
    doctorCodeUsed?: string;
    /**
     * @type string | undefined
    */
    agoraChannelName?: string;
    /**
     * @type string | undefined, date-time
    */
    createdAt?: string;
    /**
     * @type string | undefined, date-time
    */
    updatedAt?: string;
};